+++ docs/content/en/changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `toBeEndOfDay()` expectation
 - `toBeSameModelAs()` expectation
 - `toBeMidday()` expectation
+- `toBeMidnight()` expectation
 
 ### [v1.5.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.4.0...v1.5.0) - 2021-10-25